+++ nautilus_core/indicators/src/volatility/mod.rs
@@ -20,5 +20,6 @@ pub mod dc;
 pub mod fuzzy_candlesticks;
 pub mod fuzzy_enums;
 pub mod kc;
+pub mod kp;
 pub mod rvi;
 pub mod vr;